// Web/src/generated/api.d.ts
export interface components {
    schemas: {
// ...
        ExtractedSource: {
            /** Content */
            content: string;
            /** Title */
            title: string;
            /** Url */
            url?: string | null;
        };
// ...
        SourceImportView: {
            /** Id */
            id: string;
            /** Message */
            message?: string | null;
            result?: components["schemas"]["ExtractedSource"] | null;
            /**
             * Status
             * @enum {string}
             */
            status: "queued" | "processing" | "ready" | "failed";
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}